change all instances of RHFInput to Controller
