+++ src/components/GetStartedPage.tsx
@@ -17,7 +17,7 @@ import { useStateMachine } from "little-state-machine"
 import getStartedContent from "../data/getStarted"
 import TabGroup from "./TabGroup"
 import reactNative from "./codeExamples/reactNative"
-import reactNativeRHFInput from "./codeExamples/reactNativeRHFInput"
+import reactNativeController from "./codeExamples/reactNativeController"
 import typeScript from "./codeExamples/typeScript"
 import containerStyles from "../styles/container.module.css"
 import typographyStyles from "../styles/typography.module.css"
@@ -269,7 +269,7 @@ const Faq = ({
           <TabGroup buttonLabels={["Controller", "Custom Register"]}>
             <CodeArea
               isExpo
-              rawData={reactNativeRHFInput}
+              rawData={reactNativeController}
               url="https://snack.expo.io/@bluebill1049/react-hook-form-input"
             />
             <CodeArea
