+++ src/components/codeExamples/controller.ts
@@ -30,14 +30,6 @@ function App() {
         defaultValue={{ value: 'chocolate' }}
       />
       
-      {* // what is this example showing? can we remove it? *}
-      <Controller
-        as={<TextField />}
-        name="firstName"
-        control={control}
-        defaultValue=""
-      />
-      
       <button>submit</button>
     </form>
   );
